Add tidy check to deny merge commits
This will prevent users with the pre-push hook from pushing a merge
commit.

Exceptions are added for subtree updates. These exceptions are a little
hacky and may be non-exhaustive but can be extended in the future.

Checkout `master` branch in CI

Add `build_helper` crate to share code between tidy and bootstrap