Remove proc-macro back-compat hack for rental
This leaves the hack in place for allsorts-rental - it still relies
on this backc-ompat hack to compile, and some crates on crates.io
are still using it.

We've been emitting future-incompat report warnings for rental for
a while now, and fixing a broken build only requires bumping
rental to the latest point release.