include extraction in the retrying for submodule download
The code to download larger submodules previously used retries around
the `curl` invocation to handle network failures, but we saw in recent
build failures that failures can also happen during extraction, for
example if a response got terminated early.

This commit moves the retry outwards, wrapping the whole
download+extraction function in the retrying code. This means, if the
extraction fails the tarball will be re-downloaded.