Document required fd/socket flags for I/O safety.
On Unix, document that the `O_CLOEXEC` flag is required on all file
descriptors, and on Windows, document that the `WSA_FLAG_OVERLAPPED` and
`WSA_FLAG_NO_HANDLE_INHERIT` flags are required on all sockets, as a
as a requirement for I/O safety. `O_CLOEXEC` and
`WSA_FLAG_NO_HANDLE_INHERIT` prevent resources from being implicitly
leaked into child processes, and `WSA_FLAG_OVERLAPPED` prevents
undefined behavior in usage patterns which safe Rust ought to be able
to depend on.

The Rust standard library already sets these flags on all file descriptors
and sockets it creates, so there's no need to change any code in std for
this. And, this behavior is already a common expectation of crates in the
ecosystem, specifically because it avoids the problems mentioned here.

Rustix is one crate that currently creates file descriptors without
guaranteeing that `O_CLOEXEC` is set, because it was originally conveived
of as a low-level API where `O_CLOEXEC` isn't it's responsibility. However,
as a rustix author, I now think it makes sense include `O_CLOEXEC` in as a
part of the task of providing a safe API, and am submitting this PR to
seek broader consensus on this.

[rustix will follow Rust's decision here]: https://github.com/bytecodealliance/rustix/issues/495