POC: Move drop elaboration before borrowck
POC to show what's needed to move drop elaboration before borrow
checking. Ideally we could have it right aftr MIR build, but this
tries to be as minimal as possible.

There are a few hacks, mostly around drop-elaboration code that
does not pass borrowck / move analysis.

I plan to work on the remaining failing tests in the near future.